Move Python code that is necessary for running experiments into new package "testbed".
